Rewrite the referrer mapping to be more robust
The referrer code is a bit old and gnarly and only used the value of the
referrer header, but there's quite a bit of useful stuff in the query
parameters, e.g. the `utm_source` parameter.

This patch rewrites `referrers.py` to allow using this information and
get richer referrer data.